move database access
